FaceUpCardとFaceDownCardをTableCardに統合しanimationを実現
