+++ app/rooms/[roomId]/components/table/TableCardGroup.tsx
@@ -5,6 +5,7 @@ import NominateButton from './NominateButton';
 import BlankCard from './BlankCard';
 import FaceUpCard from './FaceUpCard';
 import FaceDownCard from './FaceDownCard';
+import TableCard from './TableCard';
 
 interface Props {
   player: Member;
@@ -21,10 +22,12 @@ const TableCardGroup: NextPage<Props> = ({ player, nominate }) => {
       <div className='mb-2 flex justify-center'>
         {isCardBlank ? (
           <BlankCard />
-        ) : isCardOpen ? (
-          <FaceUpCard value={player.selectedCard} />
         ) : (
-          <FaceDownCard />
+          <TableCard value={player.selectedCard} isOpen={isCardOpen} />
+          // isCardOpen ? (
+          //   <FaceUpCard value={player.selectedCard} />
+          // ) : (
+          //   <FaceDownCard />
         )}
       </div>
       <div className='text-center'>
